requirements.txt
python-telegram-bot==20.2
python-dotenv==1.0.0
google-cloud-texttospeech==2.14.1